added Label Recommendation for the service

+++ musicstore-recommendations/src/main/java/com/company/musicstorerecommendations/Repository/AlbumRecommendationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbumRecommendationRepository extends JpaRepository<AlbumRecommendation, Integer> {
+public interface AlbumRecommendationRepository extends JpaRepository<AlbumRecommendation, Long> {
 }
