+++ musicstore-recommendations/src/main/java/com/company/musicstorerecommendations/Repository/TrackRecommendationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrackRecommendationRepository extends JpaRepository<TrackRecommendation, Integer> {
+public interface TrackRecommendationRepository extends JpaRepository<TrackRecommendation, Long> {
 }
