+++ musicstore-recommendations/src/main/java/com/company/musicstorerecommendations/model/ArtistRecommendation.java
@@ -12,27 +12,27 @@ public class ArtistRecommendation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "artist_recommendation_id")
-    private Integer id;
+    private Long id;
 
     @NotNull(message = "Artist Id may not be null")
     @Column(name = "artist_id")
-    private Integer artistId;
+    private Long artistId;
 
     @NotNull(message = "User Id may not be null")
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @NotNull(message = "Liked must be true or false")
     private Boolean liked;
 
-    public ArtistRecommendation(Integer id, Integer artistId, Integer userId, Boolean liked) {
+    public ArtistRecommendation(Long id, Long artistId, Long userId, Boolean liked) {
         this.id = id;
         this.artistId = artistId;
         this.userId = userId;
         this.liked = liked;
     }
 
-    public ArtistRecommendation(Integer artistId, Integer userId, Boolean liked) {
+    public ArtistRecommendation(Long artistId, Long userId, Boolean liked) {
         this.artistId = artistId;
         this.userId = userId;
         this.liked = liked;
@@ -41,27 +41,27 @@ public class ArtistRecommendation {
     public ArtistRecommendation() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getArtistId() {
+    public Long getArtistId() {
         return artistId;
     }
 
-    public void setArtistId(Integer artistId) {
+    public void setArtistId(Long artistId) {
         this.artistId = artistId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
